Update display using flexbox
Prior to using flexbox, the display for the milk and sugar sections was
set to inline-block. This meant that those visiting the site using
screen-readers would have a difficulty accessing the web-content. The
flow of the html document would have caused confusion since both milk
and sugar sections were nested inside one section in which the content
sat beside each other on the same line.

Now, the flow of the index.html file is much easier to read since using
flexbox has given me the ability to have these two sections align
side-by-side while still following the normal flow of an html document.